backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.86.0
python-multipart==0.0.9
pydantic==2.6.0
httpx==0.28.1
supabase
razorpay
sendgrid
